Remove templated argument for sample-rate
It doesn't make any sense to generate multiple wavetables for every
possible combination of valid sample rates. The sample rate only
affects how quickly we scan through a wavetable, so let's just mandate
that when we calculate that "scan rate," that the user also pass in
the sample rate in addition to the fundamental frequency.

We'll keep the templated argument for table size, though. That seems
like a reasonable idea.