Use copy-constructors to reduce startup time
When generating complex oscillators/wavetables, reduce the amount of
work done by just copy-pasting them after one is created.

Unfortunately, the current implementation means there's a lot of
heap-allocation and calls to `new`/`delete`, which is sub-optimal and
will need to be redone in the future. But it works for now.