Move Osc 1 filter to the AudioProcessor
This makes more sense than duplicating the filter (and the atomic
pointer dereferences) for each voice in the synth. We do the filtering
once, and that's it!

Also of note: when applying the filter, it's imperative to iterate by
samples first, _then_ by channels. If you do it the other way
around (which is what I tried at first) you get crackling. I guess the
audio buffers are ordered LRLRLRLR... rather than LLLLLRRRRR...