Upload VST2 on successful CI run
Rename the CI job to just "linux".

Upload the release VST, with GCC's release build (rather than
Clang's). The debug VST is on the order of 30MB uncompressed, so we'd
like to be courteous to GitHub and upload the stripped release
VST, which is about 7MB uncompressed.

Note: I wanted to upload Clang's release build, but for whatever
reason it failed to link. Whether this is the fault of CMake or Clang
is something I'll have to investigate. For now, relegate Clang and
CMake to doing the debug build.