Don't needlessly run CI when pushing to master
We enforce that you have to make pull requests in order to get code
into master, and that the branch is up to date and passes
CI. Therefore, we don't need to run CI _again_ once the pull request
is merged. We only have 2000 minutes of CI time per month on a free
GitHub account.

We'll keep the CI cron-job that automatically runs twice a month.